增加协议
