+++ src/IoTSharp.Data.Taos/IoTSharp.Data.Taos.csproj
@@ -63,17 +63,34 @@
 	</ItemGroup>
 	<ItemGroup Condition="'$(TargetFramework)' == 'net4.6'">
 		<Reference Include="System.ComponentModel.DataAnnotations" />
+		<PackageReference Include="RestSharp">
+			<Version>106.15.0</Version>
+		</PackageReference>
+		
 	</ItemGroup>
 	<ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0' ">
 		<PackageReference Include="System.ComponentModel.Annotations" Version="5.0.0" />
-		<PackageReference Include="System.Collections.Immutable" Version="6.0.0"/>
+		<PackageReference Include="System.Collections.Immutable" Version="6.0.0" />
 	</ItemGroup>
 	<ItemGroup>
 		<PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
 	</ItemGroup>
 	<ItemGroup Condition="'$(TargetFramework)' == 'net6'">
+	  <PackageReference Include="RestSharp">
+	    <Version>108.0.3</Version>
+	  </PackageReference>
 	  <PackageReference Include="System.Collections.Immutable">
 	    <Version>6.0.0</Version>
+	  </PackageReference>
+	</ItemGroup>
+	<ItemGroup Condition="'$(TargetFramework)' == 'net7'">
+	  <PackageReference Include="RestSharp">
+	    <Version>108.0.3</Version>
+	  </PackageReference>
+	</ItemGroup>
+	<ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
+	  <PackageReference Include="RestSharp">
+	    <Version>108.0.3</Version>
 	  </PackageReference>
 	</ItemGroup>
 	<!--<ItemGroup Condition="'$(TargetFramework)' == 'net7'">
